separate script into lots of little modules
