+++ test/context.py
@@ -1,0 +1,6 @@
+import os
+import sys
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+import pasio
+__all__ = [pasio]
