+++ test/test_bench_pasio.py
@@ -1,6 +1,6 @@
 import time
 import numpy as np
-import pasio
+from .context import pasio
 import random
